requirements.txt
Django==3.0.4
gunicorn==20.0.4
django-heroku==0.3.1
google-cloud-storage==1.26.0
pillow==7.0.0